Remove toolchain auto-detection (#82)
Before this, when running cargo-bolero in a way that requires nightly, it would implicitly not use the toolchain that was explicitly provided.

For instance, if running `cargo +nightly-2022-06-20 bolero test foo`, it would still use `cargo +nightly` to compile.

This is unexpected (especially when the `nightly` toolchain is not installed under this name), and there is an easy way for the user to work around and recover the current behavior, ie. just typing `cargo +nightly`.

To the best of my understanding, just removing these two lines should be enough to come back to the behavior of "just use the caller-provided toolchain".